Don't test that "Loading..." shows
Because it is non-trivial and I'd rather keep this example simple.

The non-trivial bit stems from the fact that "Loading..." sometimes
doesn't even show up in test (as the test setup is so fast), so I'd need
to create an explicit wait middleware or something in order for it to
work consistently.